+++ cmd/issue.go
@@ -18,13 +18,8 @@
 package cmd
 
 import (
-	"fmt"
-	"log"
 	"os"
-	"strconv"
 
-	ghub "github.com/repejota/git-hub"
-	"github.com/repejota/git-hub/automation"
 	"github.com/spf13/cobra"
 )
 
@@ -40,129 +35,4 @@ var IssueCmd = &cobra.Command{
 		cmd.Usage()
 		os.Exit(0)
 	},
-}
-
-// IssueListCmd represents the issue list command
-var IssueListCmd = &cobra.Command{
-	Use:   "list",
-	Short: "List issues",
-	Long:  `List repository issues`,
-	Args:  cobra.NoArgs,
-	Run: func(cmd *cobra.Command, args []string) {
-		// Open repository
-		repo, err := ghub.OpenRepository(".")
-		if err != nil {
-			log.Fatal(err)
-		}
-
-		err = repo.GetRemoteGithubRepository("origin")
-		if err != nil {
-			log.Fatal(err)
-		}
-
-		// --repository flag
-		repository := *repo.GitHubRepository.FullName
-		if Repository != "" {
-			repository = Repository
-		}
-
-		// List issues by repo
-		issues, err := ghub.ListIssuesByRepo(repository)
-		if err != nil {
-			log.Fatal(err)
-		}
-		for _, issue := range issues {
-			fmt.Printf("#%d - %s - %s\n", *issue.Number, *issue.Title, *issue.HTMLURL)
-		}
-	},
-}
-
-// IssueStartCmd represents the issue start command
-var IssueStartCmd = &cobra.Command{
-	Use:   "start [issue number]",
-	Short: "Start an issue",
-	Long:  `Start working on an issue`,
-	Args:  cobra.ExactArgs(1),
-	Run: func(cmd *cobra.Command, args []string) {
-
-		if len(args) == 0 {
-			log.Fatalf("An issue ID is required")
-		}
-		issueID, err := strconv.Atoi(args[0])
-		if err != nil {
-			log.Fatalf("Invalid issue ID")
-		}
-		log.Printf("Start working on the issue #%d\n", issueID)
-
-		repositoryPath := "."
-
-		repository := &ghub.Repository{}
-
-		err = repository.Git(repositoryPath)
-		if err != nil {
-			log.Fatal(err)
-		}
-
-		err = repository.GetRemoteGithubRepository("origin")
-		if err != nil {
-			log.Fatal(err)
-		}
-
-		// Get User
-		user, err := ghub.GetAuthenticatedUser()
-		if err != nil {
-			log.Fatal(err)
-		}
-
-		// Get Issue
-		org, repo := ghub.ParseRepositoryFullName(*repository.GitHubRepository.FullName)
-		issue, err := ghub.GetIssue(org, repo, issueID)
-		if err != nil {
-			log.Fatal(err)
-		}
-
-		// Check if the issue is open
-		if issue.GetState() != "open" {
-			log.Printf("Issue #%d %q is %s, you can't work on it.\n", issue.GetNumber(), issue.GetTitle(), issue.GetState())
-			os.Exit(1)
-		}
-
-		// Assign User to the Issue
-		err = ghub.AssignUserToIssue(org, repo, user, issue)
-		if err != nil {
-			log.Fatal(err)
-		}
-		log.Println("Assigned issue to", user.GetLogin())
-
-		// Create local issue branch
-		issueBranchName := fmt.Sprintf("issue/%s", ghub.SlugifyIssue(issue))
-		out, err := automation.CreateLocalGitBranch(issueBranchName)
-		if err != nil {
-			log.Fatal(err)
-		}
-		log.Println("Created local branch", issueBranchName)
-		log.Println(out)
-
-		// Push local release branch to origin
-		out, err = automation.PushLocalBranchToOrigin(issueBranchName)
-		if err != nil {
-			log.Fatal(err)
-		}
-		log.Println(out)
-	},
-}
-
-// IssueFinishCmd represents the issue finish command
-var IssueFinishCmd = &cobra.Command{
-	Use:   "finish",
-	Short: "Finish an issue",
-	Long:  `Finish working on an issue`,
-	Args:  cobra.NoArgs,
-	Run: func(cmd *cobra.Command, args []string) {
-		fmt.Println("git hub issue finish")
-	},
-}
-
-func init() {
-	IssueListCmd.Flags().StringVarP(&Repository, "repository", "r", "", "Repository to get the issues from")
 }
