+++ cmd/issueList.go
@@ -21,27 +21,40 @@ import (
 	"fmt"
 	"log"
 
-	"github.com/repejota/git-hub"
+	ghub "github.com/repejota/git-hub"
 	"github.com/spf13/cobra"
 )
 
-// InfoCmd represents the info command
-var InfoCmd = &cobra.Command{
-	Use:   "info",
-	Short: "Get information about the repository",
-	Long:  `Get information about the repository and its github project`,
-	Args:  cobra.MinimumNArgs(0),
+// IssueListCmd represents the issue list command
+var IssueListCmd = &cobra.Command{
+	Use:   "list",
+	Short: "List issues",
+	Long:  `List repository issues`,
+	Args:  cobra.NoArgs,
 	Run: func(cmd *cobra.Command, args []string) {
 		// Open repository
-		repository, err := ghub.OpenRepository(".")
+		repo, err := ghub.OpenRepository(".")
 		if err != nil {
 			log.Fatal(err)
 		}
 
-		// Print info
-		fmt.Printf("Git Repository Path: %s\n", ".")
-		fmt.Printf("GitHub Repository ID: %d\n", repository.GitHubRepository.ID)
-		fmt.Printf("GitHub Repository Full Name: %s\n", *repository.GitHubRepository.FullName)
-		fmt.Printf("GitHub Repository HTML URL: %s\n", *repository.GitHubRepository.HTMLURL)
+		// --repository flag
+		repository := *repo.GitHubRepository.FullName
+		if Repository != "" {
+			repository = Repository
+		}
+
+		// List issues by repo
+		issues, err := ghub.ListIssuesByRepo(repository)
+		if err != nil {
+			log.Fatal(err)
+		}
+		for _, issue := range issues {
+			fmt.Printf("#%d - %s - %s\n", *issue.Number, *issue.Title, *issue.HTMLURL)
+		}
 	},
+}
+
+func init() {
+	IssueListCmd.Flags().StringVarP(&Repository, "repository", "r", "", "Repository to get the issues from")
 }
